fix: attempt to fix social previews not working
The meta tags are not present on the HTML returned by the server so no
image is found. I think this is because the Head component was rendered
per-page and therefore required JS to render. Here we move it to the
root _app.tsx and everything is dynamically rendered in the Head
component based on the route.