fix: attempt to fix social previews, fix build warnings (#38)
* build: dep version updates

* fix: attempt to fix social previews not working

The meta tags are not present on the HTML returned by the server so no
image is found. I think this is because the Head component was rendered
per-page and therefore required JS to render. Here we move it to the
root _app.tsx and everything is dynamically rendered in the Head
component based on the route.

* fix: fix build error regarding node vs. edge runtimes

The error was: A Node.js API is used (MessageChannel at line: 121) which
is not supported in the Edge Runtime. This was tracked down to the
headlessui Transition component in Copyable.tsx, so we remove that and
use a manual transition instead.